+++ smok/pathpoint/pathpoint.py
@@ -1,4 +1,5 @@
 import typing as tp
+import warnings
 import weakref
 from abc import ABCMeta, abstractmethod
 from concurrent.futures import Future
@@ -77,6 +78,11 @@ class Pathpoint(ReprableMixin, OmniHashableMixin, metaclass=ABCMeta):
         :param timestamp: new timestamp
         :param value: new value
         """
+        if self.current_timestamp is not None:
+            if self.current_timestamp < timestamp:
+                warnings.warn('Given lower timestamp (%s) than current one '
+                              '(%s), ignoring' % (timestamp, self.current_timestamp), UserWarning)
+                return
         self.device.pp_database.on_new_data(self.name, timestamp, value)
         if self.device.getter is not None:
             self.device.getter.data_to_update.notify()