+++ smok/threads/communicator.py
@@ -148,9 +148,15 @@ class CommunicatorThread(TerminableThread):
             self.device.api.post('/v1/device/pathpoints', json=data)
             sync.acknowledge()
         except ResponseError as e:
-            logger.debug(f'Failed to sync data', exc_info=e)
-            sync.negative_acknowledge()
-            raise
+            if e.status_code % 100 == 5:
+                logger.warning(str(data))
+                logger.debug(f'Failed to sync data', exc_info=e)
+                sync.negative_acknowledge()
+                raise
+            else:
+                logger.warning(f'Got HTTP {e.status_code} while syncing pathpoint data. '
+                               f'Assuming is it damaged and marking as synced')
+                sync.acknowledge()
 
     def sync_baob(self):
         self._sync_baob()